backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pillow==10.4.0
opencv-python==4.11.0.86
numpy==2.0.1
pytesseract==0.3.13
